Previous Work
I am unsure of when this was written, but I would guess a long time ago.
Honestly this code is just bad, and I plan to write some to a higher standard of quality soon.

+++ ProblemVII.java
@@ -14,7 +14,7 @@ public class ProblemVII {
     }
 
 
-    public static boolean isPrime(int n) {
+    public static boolean isPrime(long n) {
         //Checks if a number is negative
         if (n < 2)
             return false;
